fixed in a bug in deleting a collection and timline

+++ src/repository/timelineRepo.js
@@ -1,6 +1,8 @@
+const { collection } = require("../models/collection");
 const { Timeline, Collection } = require("../models/index");
 
 class TimelineRepo {
+  
   createtimeline = async (data, collectionId) => {
     try {
       // console.log("repositroy");
@@ -19,7 +21,14 @@ class TimelineRepo {
   };
   delete = async (id, collectionId) => {
     try {
+      const collection = await Collection.findById(collectionId);
+      if(!collection) {
+          throw new Error("Not a Valid Collection");
+      }
       const timeline = await Timeline.findByIdAndRemove(id);
+      collection.timelines = this.deleteFromArray(collection.timelines,id);
+      console.log(collection);
+      await collection.save();
       return timeline;
     } catch (error) {
       throw error;
@@ -52,5 +61,14 @@ class TimelineRepo {
       throw error;
     }
   };
+  deleteFromArray = (array, value) => {
+    let newArray = [];
+    for(let i =0;i<array.length;i++){
+      if(array[i]!=value){
+         newArray.push(array[i]);
+      }
+    }
+    return newArray;
+  }
 }
 module.exports = TimelineRepo;
